+++ KeyboardKitDemo/Shared/DemoAutocompleteSuggestionProvider.swift
@@ -3,7 +3,7 @@
 //  KeyboardKitDemo
 //
 //  Created by Daniel Saidi on 2019-07-05.
-//  Copyright © 2019 Daniel Saidi. All rights reserved.
+//  Copyright © 2021 Daniel Saidi. All rights reserved.
 //
 
 import KeyboardKit