+++ KeyboardKitDemoKeyboard/Buttons/DemoButton.swift
@@ -3,7 +3,7 @@
 //  KeyboardKitDemoKeyboard
 //
 //  Created by Daniel Saidi on 2019-04-30.
-//  Copyright © 2019 Daniel Saidi. All rights reserved.
+//  Copyright © 2021 Daniel Saidi. All rights reserved.
 //
 
 import UIKit