+++ KeyboardKitDemoKeyboard_SwiftUI/KeyboardView.swift
@@ -3,7 +3,7 @@
 //  KeyboardKitDemoKeyboard_SwiftUI
 //
 //  Created by Daniel Saidi on 2020-06-10.
-//  Copyright © 2020 Daniel Saidi. All rights reserved.
+//  Copyright © 2021 Daniel Saidi. All rights reserved.
 //
 
 import SwiftUI