+++ Sources/KeyboardKit/Input/KeyboardInputProvider.swift
@@ -3,7 +3,7 @@
 //  KeyboardKit
 //
 //  Created by Daniel Saidi on 2020-12-01.
-//  Copyright © 2020 Daniel Saidi. All rights reserved.
+//  Copyright © 2021 Daniel Saidi. All rights reserved.
 //
 
 import Foundation
@@ -16,10 +16,13 @@ import Foundation
  the input view controller's `context` when the extension is
  started. You can replace this at any time, by registering a
  new provider with the `keyboardInputProvider` property.
+
+ `IMPORTANT` This is an experimental new feature, that could
+ be redesigned in any minor release until 4.0.
  */
 public protocol KeyboardInputProvider: AnyObject {
     
-    var alphabeticInputSet: AlphabeticKeyboardInputSet { get }
-    var numericInputSet: NumericKeyboardInputSet { get }
-    var symbolicInputSet: SymbolicKeyboardInputSet { get }
+    func alphabeticInputSet(for context: KeyboardContext) -> AlphabeticKeyboardInputSet
+    func numericInputSet(for context: KeyboardContext) -> NumericKeyboardInputSet
+    func symbolicInputSet(for context: KeyboardContext) -> SymbolicKeyboardInputSet
 }