+++ Sources/KeyboardKit/Input/StaticKeyboardInputProvider.swift
@@ -3,17 +3,14 @@
 //  KeyboardKit
 //
 //  Created by Daniel Saidi on 2020-12-01.
-//  Copyright © 2020 Daniel Saidi. All rights reserved.
+//  Copyright © 2021 Daniel Saidi. All rights reserved.
 //
 
 import Foundation
 
 /**
  This keyboard input set provider returns the input set that
- it is initialized with.
- 
- It can be used to return a static input set, without taking
- factors like locales into consideration.
+ it is initialized with, regardless of the provided context.
  */
 public class StaticKeyboardInputProvider: KeyboardInputProvider {
     
@@ -25,8 +22,20 @@ public class StaticKeyboardInputProvider: KeyboardInputProvider {
         self.numericInputSet = numericInputSet
         self.symbolicInputSet = symbolicInputSet
     }
+    
+    private let alphabeticInputSet: AlphabeticKeyboardInputSet
+    private let numericInputSet: NumericKeyboardInputSet
+    private let symbolicInputSet: SymbolicKeyboardInputSet
 
-    public let alphabeticInputSet: AlphabeticKeyboardInputSet
-    public let numericInputSet: NumericKeyboardInputSet
-    public let symbolicInputSet: SymbolicKeyboardInputSet
+    public func alphabeticInputSet(for context: KeyboardContext) -> AlphabeticKeyboardInputSet {
+        alphabeticInputSet
+    }
+    
+    public func numericInputSet(for context: KeyboardContext) -> NumericKeyboardInputSet {
+        numericInputSet
+    }
+    
+    public func symbolicInputSet(for context: KeyboardContext) -> SymbolicKeyboardInputSet {
+        symbolicInputSet
+    }
 }