+++ Sources/KeyboardKit/_Deprecated/CGFloat+Deprecated.swift
@@ -3,7 +3,7 @@
 //  KeyboardKit
 //
 //  Created by Daniel Saidi on 2019-07-05.
-//  Copyright © 2019 Daniel Saidi. All rights reserved.
+//  Copyright © 2021 Daniel Saidi. All rights reserved.
 //
 
 import CoreGraphics