+++ Tests/KeyboardKitTests/Input/StandardKeyboardInputProviderTests.swift
@@ -3,7 +3,7 @@
 //  KeyboardKit
 //
 //  Created by Daniel Saidi on 2020-12-01.
-//  Copyright © 2020 Daniel Saidi. All rights reserved.
+//  Copyright © 2021 Daniel Saidi. All rights reserved.
 //
 
 import Quick
@@ -15,52 +15,71 @@ class StandardKeyboardInputProviderTests: QuickSpec {
     
     override func spec() {
         
+        var provider: KeyboardInputProvider!
+        var context: MockKeyboardContext!
+        
+        beforeEach {
+            provider = StandardKeyboardInputProvider()
+            context = MockKeyboardContext()
+        }
+        
         describe("standard keyboard input set provider") {
             
-            func provider(for identifier: String) -> KeyboardInputProvider {
-                StandardKeyboardInputProvider(locale: Locale(identifier: identifier))
+            func alphabetic(for locale: String) -> KeyboardInputSet {
+                context.locale = Locale(identifier: locale)
+                return provider.alphabeticInputSet(for: context)
+            }
+            
+            func numeric(for locale: String) -> KeyboardInputSet {
+                context.locale = Locale(identifier: locale)
+                return provider.numericInputSet(for: context)
+            }
+            
+            func symbolic(for locale: String) -> KeyboardInputSet {
+                context.locale = Locale(identifier: locale)
+                return provider.symbolicInputSet(for: context)
             }
             
             it("supports English") {
-                let prov = provider(for: "en")
-                expect(prov.alphabeticInputSet).to(equal(.alphabetic_en))
-                expect(prov.numericInputSet).to(equal(.numeric_en))
-                expect(prov.symbolicInputSet).to(equal(.symbolic_en))
+                let locale = "en"
+                expect(alphabetic(for: locale)).to(equal(.alphabetic_en))
+                expect(numeric(for: locale)).to(equal(.numeric_en))
+                expect(symbolic(for: locale)).to(equal(.symbolic_en))
             }
             
             it("supports German") {
-                let prov = provider(for: "de")
-                expect(prov.alphabeticInputSet).to(equal(.alphabetic_de))
-                expect(prov.numericInputSet).to(equal(.numeric_de))
-                expect(prov.symbolicInputSet).to(equal(.symbolic_de))
+                let locale = "de"
+                expect(alphabetic(for: locale)).to(equal(.alphabetic_de))
+                expect(numeric(for: locale)).to(equal(.numeric_de))
+                expect(symbolic(for: locale)).to(equal(.symbolic_de))
             }
             
             it("supports Italian") {
-                let prov = provider(for: "it")
-                expect(prov.alphabeticInputSet).to(equal(.alphabetic_it))
-                expect(prov.numericInputSet).to(equal(.numeric_it))
-                expect(prov.symbolicInputSet).to(equal(.symbolic_it))
+                let locale = "it"
+                expect(alphabetic(for: locale)).to(equal(.alphabetic_it))
+                expect(numeric(for: locale)).to(equal(.numeric_it))
+                expect(symbolic(for: locale)).to(equal(.symbolic_it))
             }
             
             it("supports Swedish") {
-                let prov = provider(for: "sv")
-                expect(prov.alphabeticInputSet).to(equal(.alphabetic_sv))
-                expect(prov.numericInputSet).to(equal(.numeric_sv))
-                expect(prov.symbolicInputSet).to(equal(.symbolic_sv))
+                let locale = "sv"
+                expect(alphabetic(for: locale)).to(equal(.alphabetic_sv))
+                expect(numeric(for: locale)).to(equal(.numeric_sv))
+                expect(symbolic(for: locale)).to(equal(.symbolic_sv))
             }
             
             it("has fallback support for specific locale") {
-                let prov = provider(for: "en-US")
-                expect(prov.alphabeticInputSet).to(equal(.alphabetic_en))
-                expect(prov.numericInputSet).to(equal(.numeric_en))
-                expect(prov.symbolicInputSet).to(equal(.symbolic_en))
+                let locale = "en-US"
+                expect(alphabetic(for: locale)).to(equal(.alphabetic_en))
+                expect(numeric(for: locale)).to(equal(.numeric_en))
+                expect(symbolic(for: locale)).to(equal(.symbolic_en))
             }
             
             it("has fallback support for non-existing locale") {
-                let prov = provider(for: "abcd")
-                expect(prov.alphabeticInputSet).to(equal(.alphabetic_en))
-                expect(prov.numericInputSet).to(equal(.numeric_en))
-                expect(prov.symbolicInputSet).to(equal(.symbolic_en))
+                let locale = "abc"
+                expect(alphabetic(for: locale)).to(equal(.alphabetic_en))
+                expect(numeric(for: locale)).to(equal(.numeric_en))
+                expect(symbolic(for: locale)).to(equal(.symbolic_en))
             }
         }
     }