+++ Tests/KeyboardKitTests/Mocks/MockKeyboardContext.swift
@@ -3,7 +3,7 @@
 //  KeyboardKitTests
 //
 //  Created by Daniel Saidi on 2020-07-02.
-//  Copyright © 2020 Daniel Saidi. All rights reserved.
+//  Copyright © 2021 Daniel Saidi. All rights reserved.
 //
 
 import Foundation