+++ Tests/KeyboardKitTests/UIKit/KeyboardButtonViewTests.swift
@@ -3,7 +3,7 @@
 //  KeyboardKitTests
 //
 //  Created by Daniel Saidi on 2019-05-28.
-//  Copyright © 2019 Daniel Saidi. All rights reserved.
+//  Copyright © 2021 Daniel Saidi. All rights reserved.
 //
 
 import Quick